Move App to componenets folder
