+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import booksReducer from './reducers/books';
-import App from './App';
+import App from './components/App';
 
 const initialState = {
   books: [
